Added More_external_block_finder class to find all the more external blocks of the mesh. More external blocks means blocks near the mesh boundary. These blocks could be inside/on_boundary/outside the polyhedron to discretize. This class is developed according to Single Responsibility Principle and TDD approach indeed it is added its relative tests(More_external_block_finder_tests.cpp) too.
Added Degenerate_element_finder class to find the degenerate mesh blocks/elements.This class is developed according to Single Responsibility Principle and TDD approach indeed it is added its relative tests(Degenerate_element_finder_tests.cpp) too. For the hexahedral mesh (at the moment) a degenerate element could be a pyramid, a quadrilater, a tetrehedron, a wedge, an element whose facet vertices are not coplanar, an element whose 3 vertices are collinear.